move TSNR to confounds
